+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.4.53 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Changer Event content title  for sponsors and partners
+  [thomlamb]
 
 
 1.4.52 (2026-06-23)